REFACTORED: move upload image and save/load backend things to own folders
